Fix compiler warning.
git-svn-id: https://svn.code.sf.net/p/tortoisesvn/code/trunk@21278 154be670-7969-47af-9ab0-c6fb378d2295

+++ src/TortoiseMerge/BaseView.cpp
@@ -3875,7 +3875,7 @@ bool CBaseView::HasPrevInlineDiff()
 
 void CBaseView::BeginBuildAllScreen2ViewVector()
 {
-    const int nOldSize = m_ScreenedViewLine.size();
+    const int nOldSize = (int)m_ScreenedViewLine.size();
     const int nViewCount = max(GetViewCount(), 0);
     m_ScreenedViewLine.resize(nViewCount);
     const int nCleanSize = min(nViewCount, nOldSize);
